+++ Tests/AWS.SAML/Profile/ConvertFrom-AWSCredential.Tests.ps1
@@ -7,6 +7,8 @@ Describe 'ConvertFrom-AWSCredential' {
             AccessKeyId = '123'
             SecretAccessKey = 'abc'
             SessionToken = '111'
+            AccountID = ''
+            Role = ''
             LineStart = 0
             LineEnd = 4
             TestCases = @(
@@ -18,8 +20,10 @@ Describe 'ConvertFrom-AWSCredential' {
             AccessKeyId = 'acb1d27b-de8c-48ee-98bd-0c74dda7f05'
             SecretAccessKey = '2b6fcc36-e80e-4968-82f8-b88d088741cd'
             SessionToken = '5a477223-058a-4597-ab73-0a9100efb0755a477223-058a-4597-ab73-0a9100efb0755a477223-058a-4597-ab73-0a9100efb0755a477223-058a-4597-ab73-0a9100efb0755a477223-058a-4597-ab73-0a9100efb075=='
+            AccountID = '123'
+            Role = 'user'
             LineStart = 5
-            LineEnd = 10
+            LineEnd = 13
             TestCases = @(
                 @{File = 'normal'},@{File = 'extra'},@{File = 'spacing'}
             )
@@ -29,8 +33,10 @@ Describe 'ConvertFrom-AWSCredential' {
             AccessKeyId = 'acb1d27b-de8c-48ee-98bd-0c74dda7f05'
             SecretAccessKey = ''
             SessionToken = ''
+            AccountID = ''
+            Role = ''
             LineStart = 5
-            LineEnd = 10
+            LineEnd = 13
             TestCases = @(
                @{File = 'missing'}
             )
@@ -40,8 +46,10 @@ Describe 'ConvertFrom-AWSCredential' {
             AccessKeyId = '456'
             SecretAccessKey = 'def'
             SessionToken = '222'
-            LineStart = 11
-            LineEnd = 14
+            AccountID = ''
+            Role = ''
+            LineStart = 14
+            LineEnd = 17
             TestCases = @(
                 @{File = 'normal'},@{File = 'extra'},@{File = 'spacing'},@{File = 'missing'}
             )
@@ -60,6 +68,8 @@ Describe 'ConvertFrom-AWSCredential' {
                 $response.AccessKeyId | Should Be $answer.AccessKeyId
                 $response.SecretAccessKey | Should Be $answer.SecretAccessKey
                 $response.SessionToken | Should Be $answer.SessionToken
+                $response.AccountID | Should Be $answer.AccountID
+                $response.Role | Should Be $answer.Role
             }
         }
     }
@@ -76,6 +86,8 @@ Describe 'ConvertFrom-AWSCredential' {
                 $response.AccessKeyId | Should Be $answer.AccessKeyId
                 $response.SecretAccessKey | Should Be $answer.SecretAccessKey
                 $response.SessionToken | Should Be $answer.SessionToken
+                $response.AccountID | Should Be $answer.AccountID
+                $response.Role | Should Be $answer.Role
                 $response.LineStart | Should Be $answer.LineStart
                 $response.LineEnd | Should Be $answer.LineEnd
             }