+++ Tests/AWS.SAML/Profile/Update-AWSProfile.Tests.ps1
@@ -30,7 +30,7 @@ Describe 'Update-AWSProfile' {
         }.GetNewClosure()
 
         Context "Output from Source File: [$($test.SourceFile)] matches Answer File: [$($test.AnswerFile)]" {
-            Update-AWSProfile -ProfileName 'Test' -AccessKeyId 'ff19-5d55' -SecretAccessKey '4cd2-b6da' -SessionToken '8406-ad68'
+            Update-AWSProfile -ProfileName 'Test' -AccessKeyId 'ff19-5d55' -SecretAccessKey '4cd2-b6da' -SessionToken '8406-ad68' -AccountID '123' -Role 'user'
             $answer = Get-Content -Path ".\Tests\AWS.SAML\Profile\CredentialFiles\Answer\$($test.AnswerFile)"
             $output = Get-Content -Path 'TestDrive:\output'
             
@@ -38,11 +38,33 @@ Describe 'Update-AWSProfile' {
                 Assert-MockCalled Save-AWSCredentialFile -ModuleName AWS.SAML.Profile -Exactly -Times 1 -Scope Context
             }
 
+            It 'File is same length' {
+                $output.Count | Should Be $answer.Count
+            }
+
             for ($i = 0; $i -lt $answer.Count; $i++) {
                 It "Line $($i+1) - matches" {
                     $output[$i] | Should Be $answer[$i]
                 } 
             }
+        }
+    }
+
+    Context 'Errors' {
+        it 'When File is blank' {
+            Mock Get-AWSCredentialFile -ModuleName AWS.SAML.Profile { 
+                return $null
+            }
+
+            {Update-AWSProfile -ProfileName 'Test' -AccessKeyId 'ff19-5d55' -SecretAccessKey '4cd2-b6da' -SessionToken '8406-ad68' -AccountID '123' -Role 'user'} | Should -Throw
+        }
+
+        it 'Profile not found' {
+            Mock Get-AWSCredentialFile -ModuleName AWS.SAML.Profile { 
+                return '[Dev]'
+            }
+
+            {Update-AWSProfile -ProfileName 'Test' -AccessKeyId 'ff19-5d55' -SecretAccessKey '4cd2-b6da' -SessionToken '8406-ad68' -AccountID '123' -Role 'user'} | Should -Throw
         }
     }
 }