+++ Tests/AWS.SAML/Settings/Get-AWSCredentialFile.Tests.ps1
@@ -14,8 +14,8 @@ Describe 'Get-AWSCredentialFile' {
             Assert-MockCalled Get-AWSDirectory -ModuleName AWS.SAML.Settings -Exactly -Times 1
         }
 
-        It 'Returns empty string' {
-            $response | Should Be ''
+        It 'Returns null'{
+            $response | Should Be $null
         }
     }
 
@@ -27,12 +27,29 @@ Describe 'Get-AWSCredentialFile' {
             Assert-MockCalled Get-AWSDirectory -ModuleName AWS.SAML.Settings -Exactly -Times 1
         }
 
-        It 'Returns empty string' {
-            $response | Should Be ''
+        It 'Returns null'{
+            $response | Should Be $null
         }
     }
 
-    Context 'File present' {
+    Context 'Empty file' {
+        
+        # Create File
+        New-Item -ItemType Directory $dir
+        New-Item -ItemType File "$dir`credentials"
+
+        $response = Get-AWSCredentialFile
+
+        It 'Calls Get-AWSDirectory once' {
+            Assert-MockCalled Get-AWSDirectory -ModuleName AWS.SAML.Settings -Exactly -Times 1
+        }
+
+        It 'Returns null'{
+            $response | Should Be $null
+        }
+    }
+
+    Context 'File with content' {
         $content = @('abc','123')
         
         # Create File
@@ -44,6 +61,10 @@ Describe 'Get-AWSCredentialFile' {
 
         It 'Calls Get-AWSDirectory once' {
             Assert-MockCalled Get-AWSDirectory -ModuleName AWS.SAML.Settings -Exactly -Times 1
+        }
+
+        It 'Returns an array'{
+            $response.GetType().BaseType.Name | Should Be 'Array'
         }
 
         It 'Returns file content' {